View -> Slice/AsTuple/AsDict
