+++ chainer/dataset/tabular/as_mode.py
@@ -1,0 +1,41 @@
+from chainer.dataset import TabularDataset
+
+
+class AsTuple(TabularDataset):
+
+    def __init__(self, dataset):
+        self._dataset = dataset
+
+    def __len__(self):
+        return len(self._dataset)
+
+    @property
+    def keys(self):
+        return self._dataset.keys
+
+    @property
+    def mode(self):
+        return tuple
+
+    def get_examples(self, indices, key_indices):
+        return self._dataset.get_examples(indices, key_indices)
+
+
+class AsDict(TabularDataset):
+
+    def __init__(self, dataset):
+        self._dataset = dataset
+
+    def __len__(self):
+        return len(self._dataset)
+
+    @property
+    def keys(self):
+        return self._dataset.keys
+
+    @property
+    def mode(self):
+        return dict
+
+    def get_examples(self, indices, key_indices):
+        return self._dataset.get_examples(indices, key_indices)
