+++ tests/chainer_tests/dataset_tests/tabular_tests/test_as_mode.py
@@ -1,0 +1,38 @@
+import unittest
+
+from chainer import testing
+from chainer.dataset import TabularDataset
+
+from .test_tabular_dataset import DummyDataset
+
+
+@testing.parameterize(
+    {'mode': tuple},
+    {'mode': dict},
+)
+class TestAsTuple(unittest.TestCase):
+
+    def test_as_tuple(self):
+        dataset = DummyDataset(self.mode, None)
+        view = dataset.as_tuple()
+        self.assertIsInstance(view, TabularDataset)
+        self.assertEqual(len(view), len(dataset))
+        self.assertEqual(view.keys, dataset.keys)
+        self.assertEqual(view.mode, tuple)
+
+
+@testing.parameterize(
+    {'mode': tuple},
+    {'mode': dict},
+)
+class TestAsDict(unittest.TestCase):
+
+    def test_as_dict(self):
+        dataset = DummyDataset(self.mode, None)
+        view = dataset.as_dict()
+        self.assertIsInstance(view, TabularDataset)
+        self.assertEqual(len(view), len(dataset))
+        self.assertEqual(view.keys, dataset.keys)
+
+
+testing.run_module(__name__, __file__)
